Complete rewrite
The first iteration of dina-schema was a proof of concept. This rewrite
is a cleanup and has a structure better prepared for what we want to do:
IT:
- Read and combine files in one step.
- Result from read is used to create swagger spec, openApi spec and
models
Also:
- Paths are renamed to endpoints (might later be renamed to operations)
and now we keep endpoint specific models like responses and requests
here instead of in a centralised location.
- Its now possible to specify parameters as subfolders in the models
and add description in markdown. For now these are merged into the
models but might later live as separate entities
- Now we are testing the resulting swagger and openApi specification
against swagger and openApi json schema

Finally: A lot of code have been moved around so it will make sense to
test both the swagge and openApi documentation and dina-collections-ui